miniProject2/requirements.txt
Django>=4.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
pymongo>=4.6.0
mysqlclient>=2.2.0
celery>=5.3.6
redis>=5.0.1
spacy>=3.7.2
python-docx>=1.0.1
PyPDF2>=3.0.1
pydantic>=2.5.2
drf-yasg>=1.21.7
sentry-sdk>=1.39.1
python-dotenv>=1.0.0
django-filter>=23.5
django-redis>=5.4.0
django-ratelimit>=4.1.0
nltk>=3.8.1
djongo>=1.3.6
marshmallow>=3.20.1
Pillow>=10.1.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-magic>=0.4.27
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
gunicorn>=21.2.0
whitenoise>=6.5.0
django-storages>=1.14.0
boto3>=1.28.0 